Add mlabutils.config module and a test for it.

+++ src/mlabutils/fileman.py
@@ -18,25 +18,25 @@ class FileEntry(object):
         self.rel_path = rel_path
 
         self.variables = {}
-    
+
     def __repr__(self):
         return "FileEntry(%r, %r)" % (self.directory, self.rel_path, )
-    
+
     @property
     def description(self):
         return "file %s in %s" % (
             self.rel_path,
             self.directory,
         )
-    
+
     def set_var(self, name, value):
         self.variables[str(name)] = value
-    
+
     def iter_dir(self):
         for f in os.listdir(self.abs_path):
             entry = FileEntry(self.directory, path.join(self.rel_path, f))
             yield entry
-    
+
     @property
     def abs_path(self):
         return path.abspath(path.join(self.directory, self.rel_path))
@@ -48,11 +48,15 @@ class FileEntry(object):
     @property
     def dirname(self):
         return path.dirname(self.abs_path)
-    
+
+    @property
+    def exists(self):
+        return path.exists(self.abs_path)
+
     @property
     def is_file(self):
         return path.isfile(self.abs_path)
-    
+
     @property
     def is_dir(self):
         return path.isdir(self.abs_path)
@@ -64,47 +68,47 @@ class FileEntry(object):
         )
 
 
-class FileEntryStream(object):
-    def __init__(self, source, filters = None):
-        self.source = source
-        self.filters = [] if filters is None else list(filters)
-
-    @property
-    def description(self):
-        result = self.source.description
-        for filter in self.filters:
-            result += " " + filter.description
-        return result
-    
-    def filter(self, filter):
-        return FileEntryStream(self.source, self.filters + [filter, ])
-    
-    def __iter__(self):
-        for entry in self.source:
-            filtered = entry
-            for filter in self.filters:
-                filtered = filter(filtered)
-                if filtered is None:
-                    break
-            if filtered is None:
-                continue
-            yield filtered
-
-
 class FileSource(object):
     def __iter__(self):
         return self.iter_files()
-    
+
     @property
     def description(self):
         return "unknown file source"
-    
+
     def iter_files(self):
         if False:
             yield None
-    
+
     def filter(self, filter):
-        return FileEntryStream(self, [filter, ])
+        return FilteredFileSource(self, filter)
+
+
+class FilteredFileSource(FileSource):
+    def __init__(self, source, filter):
+        FileSource.__init__(self)
+        self.source = source
+        self.filter = filter
+
+    @property
+    def description(self):
+        return self.source.description + " " + self.filter.description
+
+    def iter_files(self):
+        for entry in self.source:
+            if entry is None:
+                continue
+
+            filtered = self.filter(entry)
+
+            if filtered is None:
+                continue
+            elif isinstance(filtered, FileEntry):
+                yield filtered
+            else:
+                for item in filtered:
+                    if item is not None:
+                        yield item
 
 
 class DirectoryFileSource(FileSource):
@@ -112,36 +116,36 @@ class DirectoryFileSource(FileSource):
         FileSource.__init__(self)
         self.directory = directory
         self.recursive = recursive
-    
+
     @property
     def description(self):
         return "files in directory %s" % (path.abspath(self.directory), )
-    
+
     def iter_files(self):
         if not path.isdir(self.directory):
             return
-        
+
         abs_dir = path.abspath(self.directory)
         queue = [FileEntry(abs_dir, "."), ]
-        
+
         while len(queue) > 0:
             current = queue.pop()
-            
+
             for entry in current.iter_dir():
                 yield entry
-                
+
                 if self.recursive and entry.is_dir:
                     queue.append(entry)
-        
+
 
 class FileFilter(object):
     @property
     def description(self):
         return "unknown file filter"
-    
+
     def __call__(self, entry, *args, **kwargs):
         return self.apply(entry, *args, **kwargs)
-    
+
     def apply(self, entry):
         return entry
 
@@ -151,11 +155,11 @@ class RegExFilter(FileFilter):
         FileFilter.__init__(self)
         self.pattern = re.compile(pattern)
         self.set_variables = set_variables
-    
+
     @property
     def description(self):
         return "matching pattern %r" % (self.pattern.pattern, )
-    
+
     def apply(self, entry):
         match = self.pattern.match(entry.basename)
         if match is None:
@@ -170,7 +174,7 @@ class IsFileFilter(FileFilter):
     @property
     def description(self):
         return "which are files"
-    
+
     def apply(self, entry):
         if not entry.is_file:
             return None
@@ -185,7 +189,7 @@ class FormatFilter(FileFilter):
     @property
     def description(self):
         return "with name transformed to %r" % (self.format, )
-    
+
     def apply(self, entry):
         variables = dict(entry.variables)
         variables["rel_path"] = path.dirname(entry.rel_path)
@@ -198,7 +202,7 @@ class FormatFilter(FileFilter):
 class OlderThanFilter(FileFilter):
     def __init__(self, ):
         self.descending = descending
-    
+
     def apply(self, entry):
         pass
 
@@ -211,9 +215,11 @@ class FileTask(object):
 class DeleteTask(FileTask):
     def __init__(self, remove_empty_parents = False):
         FileTask.__init__(self)
-    
+
     def execute(self, files):
         for f in files:
+            if not f.exists:
+                continue
             if f.is_file:
                 os.remove(f.abs_path)
             elif f.is_dir:
@@ -222,6 +228,11 @@ class DeleteTask(FileTask):
                 os.removedirs(f.dirname)
             except OSError:
                 pass
+
+
+class SyncTask(FileTask):
+    def execute(self, files):
+        pass
 
 
 def main():
